Re-restructure New Data Classes
