+++ rfsocinterface/core/data/map.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 from pathlib import Path
 import pdb
 
-import h5py
 import numpy as np
 import numpy.typing as npt
 from scipy import signal, ndimage
@@ -13,9 +12,8 @@ import matplotlib.pyplot as plt
 from matplotlib.backends.backend_pdf import PdfPages
 from kidpy3 import RawDataFile
 
-from rfsocinterface.core.data.routines import DataRoutine, Mapper
-from rfsocinterface.core.utils import gaussian_filter, GAUSSIAN_SIGMA
-from rfsocinterface.core.data import N_POLARIZATION, ProcessedData, PyTablesProcessedData, PyTablesMapData, MapData, remove_electronics_noise, rotate_basis, generate_calibrated_data
+from rfsocinterface.core.data.routines import CleanTOD, DataRoutine, Downsample, HighPassFilter, LowPassFilter, Mapper
+from rfsocinterface.core.data import N_POLARIZATION, ProcessedData, PyTablesProcessedData, PyTablesMapData, MapData
 
 DECIMATE_ORDER = 5
 BUTTER_ORDER = 6
@@ -37,71 +35,6 @@ def get_map_size(map: MapData, az_trim: float, za_trim: float, map_dpix: float, 
         map_y += 0.1  # 0.1 accounts for assymmetry in array
 
     return n_pix_x, n_pix_y, map_x, map_y
-
-
-class RemoveElectronicsNoise(DataRoutine):
-    def __init__(self):
-        super().__init__()
-    
-    def forward(self, pd: ProcessedData) -> ProcessedData:
-        gain_phase_data = pd.data_gain_phase
-        clean_gain_phase_data = remove_electronics_noise(gain_phase_data)
-
-        new_data_freq_diss, new_data_mK = generate_calibrated_data(
-            clean_gain_phase_data,
-            pd.IQ_to_gain_phase_angle,
-            pd.dIQ_df,
-            pd.df_per_mK
-        )
-        return pd.with_values(
-            data_gain_phase=clean_gain_phase_data,
-            data_freq_diss=new_data_freq_diss,
-            data_mK=new_data_mK,
-        )
-
-
-class CleanTOD(DataRoutine):
-
-    def __init__(
-            self,
-            save_file: bool=True,
-    ):
-        super().__init__()
-        self.save_file = save_file
-
-    def forward(self, md: MapData) -> MapData:
-
-        if not isinstance(md, MapData):
-            md = MapData.from_processed_data(md)
-        data = md.data_mK
-        chanmask = md.chanmask
-        data_clean = np.copy(data)
-        good_samples = md.get_good_samples()
-        
-        #average template subtraction
-        goodchan = np.ndarray.flatten(np.argwhere(chanmask == 1))
-        # pdb.set_trace()
-        data_good = data[goodchan][:, good_samples]
-        template = np.sum(data_good, axis=0)
-        template = template - np.mean(template)
-        template_corr = np.sum(np.multiply(data_good,template), axis=1) / \
-                        np.sum(np.multiply(template,template))
-        data_clean_good = data_good - np.outer(template_corr, template)
-        data_clean[goodchan][:, good_samples] = data_clean_good
-
-        if self.save_file:
-            with h5py.File(md.cleaned_file_template, 'w') as cfile:
-                cfile.create_dataset("chanmask", data=chanmask)
-                cfile.create_dataset("detector_pol", data=md.detector_pol)
-                cfile.create_dataset("clean_data", data=data_clean)
-                cfile.create_dataset("time", data=md.timestamp)
-                cfile.create_dataset("detector_az", data=md.detector_az)    
-                cfile.create_dataset("detector_za", data=md.detector_za)
-
-
-        return md.with_values(
-            data_mK=data_clean,
-        )
 
 
 class RemovePointLomaPickup(DataRoutine):
@@ -264,73 +197,6 @@ class BinTODIntoMap(DataRoutine):
             chanmask=new_chanmask,
         )
             
-class GaussianFilter(DataRoutine):
-    def __init__(self, gaussian_sigma: tuple[float, float]=GAUSSIAN_SIGMA):
-        super().__init__()
-        self.gaussian_sigma = gaussian_sigma
-    
-    def forward(self, pd: ProcessedData, field: str='data_mK') -> ProcessedData:
-        smoothed_data = gaussian_filter(pd.__getattribute__(field), self.gaussian_sigma)
-        return pd.with_values(**{field: smoothed_data})
-
-
-class CutoffFilter(DataRoutine):
-    def __init__(self, filter_freq: float, btype: str):
-        super().__init__()
-        self.filter_freq = filter_freq
-        self.btype = btype
-
-    def forward(self, pd: ProcessedData) -> ProcessedData:
-        filt_sos = signal.butter(BUTTER_ORDER, self.filter_freq, btype=self.btype, fs=pd.fs, output='sos', analog=False)
-
-        # Apply cutoff filter
-        data_gain_phase_filt = signal.sosfiltfilt(filt_sos, pd.data_gain_phase)
-        data_freq_diss_filt = signal.sosfiltfilt(filt_sos, pd.data_freq_diss)
-        data_mK_filt = signal.sosfiltfilt(filt_sos, pd.data_mK)
-        return pd.with_values(
-            data_gain_phase=data_gain_phase_filt,
-            data_freq_diss=data_freq_diss_filt,
-            data_mK=data_mK_filt,
-        )
-
-class LowPassFilter(CutoffFilter):
-    def __init__(self, filter_freq: float):
-        super().__init__(filter_freq, btype='lowpass')
-
-
-class HighPassFilter(CutoffFilter):
-    def __init__(self, filter_freq: float):
-        super().__init__(filter_freq, btype='highpass')
-
-
-class Downsample(DataRoutine):
-    def __init__(self, ds_factor: float=6, order: int=DECIMATE_ORDER):
-        super().__init__()
-        self.ds_factor = ds_factor
-        self.order=order
-    
-    def forward(self, pd: ProcessedData) -> ProcessedData:
-        data_freq_diss_ds = signal.decimate(pd.data_freq_diss, self.ds_factor)
-        data_gain_phase_ds = signal.decimate(pd.data_gain_phase, self.ds_factor)
-        data_mK_ds = signal.decimate(pd.data_mK, self.ds_factor)
-        timestamp_ds = signal.decimate(pd.timestamp, self.ds_factor)
-        if np.size(pd.detector_az) > 1:
-            detector_az_ds = signal.decimate(pd.detector_az, self.ds_factor, n=self.order, axis=1)
-            detector_za_ds = signal.decimate(pd.detector_za, self.ds_factor, n=self.order, axis=1)
-        else:
-            detector_az_ds = pd.detector_az
-            detector_za_ds = pd.detector_za
-        return pd.with_values(
-            data_freq_diss=data_freq_diss_ds,
-            data_gain_phase=data_gain_phase_ds,
-            data_mK=data_mK_ds,
-            timestamp=timestamp_ds,
-            detector_az=detector_az_ds,
-            detector_za=detector_za_ds,
-        )
-
-
-
 class BasicMapRemoval(DataRoutine):
     def forward(self, map: MapData) -> MapData:
         map_data = map.data_mK
